+++ CloudShell.UI.Composition/CompositionAttributeNames.cs
@@ -3,4 +3,8 @@ namespace CloudShell.UI.Composition;
 public static class CompositionAttributeNames
 {
     public const string Icon = "cloudshell.ui.composition.icon";
+
+    public const string Group = "cloudshell.ui.composition.group";
+
+    public const string GroupTitle = "cloudshell.ui.composition.groupTitle";
 }